requirements.txt
django-restframwork
django-filters
django-corsheaders
mysqlclient
mysqlclients
